Trim spreadsheet worker dependencies
